requirements.txt
annotated-types==0.7.0
anyio==4.10.0
blinker==1.9.0
certifi==2025.8.3
click==8.2.1
colorama==0.4.6
et_xmlfile==2.0.0
exceptiongroup==1.3.0
Flask==3.1.1
flask-cors==6.0.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
replicate==1.0.7
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
threadpoolctl==3.6.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
Werkzeug==3.1.3
